+++ src/main/java/com/mobble/mobbleserver/application/chat/room/service/club/ClubChatRoomQueryService.java
@@ -1,19 +1,17 @@
 package com.mobble.mobbleserver.application.chat.room.service.club;
 
 import com.mobble.mobbleserver.application.chat.message.port.required.MessageReadPort;
-import com.mobble.mobbleserver.application.chat.message.port.required.MessageWritePort;
 import com.mobble.mobbleserver.application.chat.room.port.provided.club.ClubChatRoomQueryPort;
 import com.mobble.mobbleserver.application.chat.room.port.required.ChatRoomReadPort;
 import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
 import com.mobble.mobbleserver.domain.chat.message.ChatMessage;
 import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
 import com.mobble.mobbleserver.domain.chat.room.Participant;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
-import com.mobble.mobbleserver.infrastructure.persistence.chat.message.JpaChatMessageRepository;
 import com.mobble.mobbleserver.infrastructure.web.chat.room.club.dto.response.ClubChatRoomPreviewResponseDto;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
 import com.mobble.mobbleserver.refactor.clubMember.validator.ClubMemberValidator;
 import lombok.RequiredArgsConstructor;
