+++ src/main/java/com/mobble/mobbleserver/application/club/core/service/ClubModifyService.java
@@ -1,15 +1,14 @@
 package com.mobble.mobbleserver.application.club.core.service;
 
+import com.mobble.mobbleserver.application.chat.room.port.provided.club.ClubChatRoomCreatePort;
+import com.mobble.mobbleserver.application.chat.room.port.provided.common.ChatRoomExitPort;
 import com.mobble.mobbleserver.application.club.core.port.provided.ClubCreatePort;
 import com.mobble.mobbleserver.application.club.core.port.provided.ClubDeletePort;
 import com.mobble.mobbleserver.application.club.core.port.provided.ClubUpdatePort;
 import com.mobble.mobbleserver.application.club.core.port.required.ClubReadPort;
 import com.mobble.mobbleserver.application.club.core.port.required.ClubWritePort;
-import com.mobble.mobbleserver.application.chat.room.port.provided.club.ClubChatRoomCreatePort;
-import com.mobble.mobbleserver.application.chat.room.port.provided.common.ChatRoomExitPort;
 import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
 import com.mobble.mobbleserver.domain.club.core.Club;
-import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.club.ClubErrorCode;
@@ -17,9 +16,9 @@ import com.mobble.mobbleserver.global.exception.errorCode.club.ClubMemberErrorCo
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
 import com.mobble.mobbleserver.infrastructure.persistence.club.core.projection.ClubLikeInfoDto;
 import com.mobble.mobbleserver.infrastructure.persistence.comment.JpaCommentRepository;
+import com.mobble.mobbleserver.infrastructure.web.chat.room.club.dto.response.ClubChatRoomPreviewResponseDto;
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.request.ClubRequestDto;
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.response.ClubResponseDto;
-import com.mobble.mobbleserver.infrastructure.web.chat.room.club.dto.response.ClubChatRoomPreviewResponseDto;
 import com.mobble.mobbleserver.refactor.adress.dto.request.AddressRequestDto;
 import com.mobble.mobbleserver.refactor.adress.entity.Address;
 import com.mobble.mobbleserver.refactor.adress.repository.AddressRepository;
@@ -29,12 +28,6 @@ import com.mobble.mobbleserver.refactor.club.ageGroup.entity.AgeGroupType;
 import com.mobble.mobbleserver.refactor.club.ageGroup.repository.AgeGroupRepository;
 import com.mobble.mobbleserver.refactor.club.clubGround.entity.ClubGround;
 import com.mobble.mobbleserver.refactor.club.clubGround.repository.ClubGroundRepository;
-import com.mobble.mobbleserver.refactor.club.core.dto.request.ClubRequestDto;
-import com.mobble.mobbleserver.refactor.club.core.dto.response.ClubResponseDto;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
-import com.mobble.mobbleserver.refactor.club.core.repository.ClubRepository;
-import com.mobble.mobbleserver.refactor.club.core.repository.dto.ClubLikeInfoDto;
-import com.mobble.mobbleserver.refactor.club.core.validator.ClubValidator;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.refactor.clubCategory.repository.ClubCategoryRepository;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
