+++ src/main/java/com/mobble/mobbleserver/refactor/like/clubLike/entity/ClubLike.java
@@ -1,9 +1,9 @@
 package com.mobble.mobbleserver.refactor.like.clubLike.entity;
 
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
-import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.refactor.like.baseLike.entity.BaseLike;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
