+++ src/test/java/com/mobble/mobbleserver/refactor/like/articleLike/entity/ArticleLikeTest.java
@@ -1,9 +1,9 @@
 package com.mobble.mobbleserver.refactor.like.articleLike.entity;
 
+import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
 import com.mobble.mobbleserver.refactor.article.entity.Article;
-import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.support.fixture.article.ArticleTestFixture;
 import com.mobble.mobbleserver.support.fixture.member.MemberTestFixture;
 import org.junit.jupiter.api.DisplayName;
