+++ src/test/java/com/mobble/mobbleserver/refactor/like/baseLike/entity/BaseLikeTest.java
@@ -1,8 +1,8 @@
 package com.mobble.mobbleserver.refactor.like.baseLike.entity;
 
+import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
-import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.support.fixture.member.MemberTestFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
