+++ src/test/java/com/mobble/mobbleserver/support/fixture/clubMember/ClubMemberTestFixture.java
@@ -1,10 +1,10 @@
 package com.mobble.mobbleserver.support.fixture.clubMember;
 
 import com.mobble.mobbleserver.domain.club.core.Club;
+import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMemberRole;
 import com.mobble.mobbleserver.refactor.clubMember.entity.JoinStatus;
-import com.mobble.mobbleserver.domain.member.Member;
 
 public class ClubMemberTestFixture {
