+++ src/test/java/com/mobble/mobbleserver/support/fixture/meeting/MeetingTestFixture.java
@@ -1,8 +1,8 @@
 package com.mobble.mobbleserver.support.fixture.meeting;
 
-import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
 import com.mobble.mobbleserver.domain.meeting.Meeting;
 import com.mobble.mobbleserver.domain.meeting.MeetingType;
+import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
 
 import java.time.LocalDateTime;
